feat: use game_start_date for registration and start flags

+++ scripts/add_config.js
@@ -10,16 +10,19 @@ const Config = require('../models/config');
 const game_start_date_cmd = process.argv[2];
 
 (async () => {
+  if (!game_start_date_cmd) {
+    console.error('Please provide game start date!');
+    process.exit(1);
+  }
   const session = await mongoose.startSession();
   try {
+    const game_start_date = new Date(game_start_date_cmd);
     await session.withTransaction(async () => {
       await Config.deleteOne({ id: 0 });
       await Config.create({
         is_game_over: false,
-        is_game_started: false,
-        is_register_closed: false,
         ...(game_start_date_cmd && {
-          game_start_date: new Date(game_start_date_cmd)
+          game_start_date
         }),
         id: 0
       });
